Styled the graphs section with flex-box
I made the display of graphs-section a flex box so that each direct child of that div is a flex-item. This makes it easier to account for smaller screen sizes and to set the three column layout shown in the screenshot. I added some margins to the image as well and increased the font size of the text to best match the screenshot