Write JSX for the bottom section
The bottom section contains a header, 3 columns, each of which contain a sub-heading and a paragraph, and finally, a button to 'read more'. The three columns are wrapped in a container and each individual column is also wrapped in its own div tag. I assigned class names according to the BEM naming convention