Styled the bottom section.
I increased the font-size for the h2 and h3 sub-headings. I made the bottom-section__content a flex-box and each item in it have a flex basis of 25%. I also set the flex-box to justify content with space around, simulating the margins showed in the screen shot. Lastly, I added margins to top and bottom of bottom-section to best match what is shown in the screenshot.