Added some conditional statements and refactored the code
The data from the API is shaped differently then the data from test_data.json. For starters, it returns an object, not a list. That's why I removed all the data[0]['yearResults'] and replaced it with data['yearResults']. Secondly, the citizenship summary does not have a 'yearResults' attribute, so I had to add conditional statements to prevent an error from occuring.