Changed the code so it's compatible with the new data format.
In the test data, citizenshipResults is an key in the 'first' element of the array, so that's why there was a whole bunch of data[0]['citizinshipResults']. This is not the case for the data from the api. Instead, it's an array of objects, each object containing data for each country. So instead, the call should just be 'data'.